feat: 패키지 구조 변경 - 패키지 병합, 구조변경으로 인한 빠른 배포 type: short summary message body
