--- src/main/java/mandykr/nutrient/entity/Combination.java
@@ -1,47 +1,0 @@
-package mandykr.nutrient.entity;
-
-import lombok.*;
-import mandykr.nutrient.entity.util.BaseTimeEntity;
-
-import javax.persistence.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static javax.persistence.FetchType.LAZY;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@EqualsAndHashCode(of = "id")
-public class Combination extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "COMBINATION_ID")
-    private Long id;
-
-    private String caption;
-
-    private Double rating;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "MEMBER_ID")
-    private Member member;
-
-    @OneToMany(mappedBy = "combination")
-    @Builder.Default
-    private List<CombinationStarRate> combinationStarRates = new ArrayList<>();
-
-    public void updateRating() {
-        this.rating = combinationStarRates.stream().mapToInt(o -> o.getStarNumber()).average().getAsDouble();
-    }
-
-    public void updateList(CombinationStarRate combinationStarRate) {
-        int idx = combinationStarRates.indexOf(combinationStarRate);
-        combinationStarRates.set(idx, combinationStarRate);
-    }
-
-}
