+++ src/main/java/mandykr/nutrient/service/CombinationStarRateService.java
@@ -2,11 +2,11 @@ package mandykr.nutrient.service;
 
 import lombok.RequiredArgsConstructor;
 import mandykr.nutrient.dto.request.CombineStarRateRequest;
-import mandykr.nutrient.entity.Combination;
-import mandykr.nutrient.entity.CombinationStarRate;
+import mandykr.nutrient.entity.combination.Combination;
+import mandykr.nutrient.entity.combination.CombinationStarRate;
 import mandykr.nutrient.entity.Member;
-import mandykr.nutrient.repository.CombinationRepository;
-import mandykr.nutrient.repository.CombinationStarRateRepository;
+import mandykr.nutrient.repository.combination.starrate.CombinationStarRateRepository;
+import mandykr.nutrient.repository.combination.CombinationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
